feat: provide a way to filter flags
So some command like kubectl use flags that both end or don't with an
`=` char, pcomplete only supports one or the other, so we have to
strip all the flags that end with `=`.

But it gets more complicated, because some kubectl commands only have
auto-completion for flags ending in an `=`.  It's just noise in the
data.  So for those we need to convert them into flags without an
equals.